backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
langchain-core
httpx
supabase
python-dotenv
apscheduler
python-dotenv